+++ plugins/ops/apps-visualizer/test/integration/conftest.py
@@ -1,7 +1,9 @@
+import asyncio
 import json
 import os
 from pathlib import Path
 from datetime import datetime
+import socket
 
 import pytest
 
@@ -10,62 +12,10 @@ from hopeit.server import config as server_config
 from hopeit.testing.apps import config, execute_event
 
 from hopeit.log_streamer import LogRawBatch, LogBatch
-from hopeit.config_manager.runtime import runtime
 
-from . import MockServer
-from hopeit.config_manager import RuntimeAppInfo, RuntimeApps, ServerInfo, ServerStatus
-import socket
+from . import mock_getenv
 
-
-def mock_getenv(var_name):
-    if var_name == "HOPEIT_APPS_VISUALIZER_HOSTS":
-        return "in-process"
-    if var_name == "HOPEIT_SIMPLE_EXAMPLE_HOSTS":
-        return "test-host"
-    if var_name == "HOPEIT_APPS_API_VERSION":
-        return APPS_API_VERSION
-    if var_name == "HOPEIT_APPS_ROUTE_VERSION":
-        return APPS_ROUTE_VERSION
-    raise NotImplementedError(var_name)
-
-
-@pytest.fixture
-def runtime_apps(monkeypatch):
-    monkeypatch.setattr(server_config.os, 'getenv', mock_getenv)
-    app_config = config('apps/examples/simple-example/config/app-config.json')
-    client_app_config = config('apps/examples/client-example/config/app-config.json')
-    monkeypatch.setattr(
-        runtime,
-        "server",
-        MockServer(app_config, client_app_config)
-    )
-    return RuntimeApps(
-        apps={
-            app_config.app_key(): RuntimeAppInfo(
-                servers=[
-                    ServerInfo(
-                        host_name=socket.gethostname(),
-                        pid=str(os.getpid()),
-                        url="in-process"
-                    )
-                ],
-                app_config=app_config
-            ),
-            client_app_config.app_key(): RuntimeAppInfo(
-                servers=[
-                    ServerInfo(
-                        host_name=socket.gethostname(),
-                        pid=str(os.getpid()),
-                        url="in-process"
-                    )
-                ],
-                app_config=client_app_config
-            )
-        },
-        server_status={
-            "in-process": ServerStatus.ALIVE
-        }
-    )
+_mock_lock = asyncio.Lock()
 
 
 @pytest.fixture
@@ -75,8 +25,15 @@ def plugin_config(monkeypatch):
 
 
 @pytest.fixture
-def events_graph_data():
-    with open(Path(os.path.dirname(os.path.realpath(__file__))) / 'events_graph_data.json') as f:
+def events_graph_data_standard():
+    with open(Path(os.path.dirname(os.path.realpath(__file__))) / 'events_graph_data_standard.json') as f:
+        json_str = f.read().replace("${APPS_ROUTE_VERSION}", APPS_ROUTE_VERSION)
+        return json.loads(json_str)
+
+
+@pytest.fixture
+def events_graph_data_expanded():
+    with open(Path(os.path.dirname(os.path.realpath(__file__))) / 'events_graph_data_expanded.json') as f:
         json_str = f.read().replace("${APPS_ROUTE_VERSION}", APPS_ROUTE_VERSION)
         return json.loads(json_str)
 
@@ -84,6 +41,11 @@ def events_graph_data():
 async def _process_log_entries(raw: LogRawBatch) -> LogBatch:
     plugin_config = config('plugins/ops/log-streamer/config/plugin-config.json')
     return await execute_event(plugin_config, "log_reader", payload=raw)  # type: ignore
+
+
+@pytest.fixture
+def mock_lock():
+    return _mock_lock
 
 
 @pytest.fixture