+++ plugins/ops/config-manager/test/integration/test_runtime_apps_config.py
@@ -17,4 +17,23 @@ async def test_runtime_apps_config(monkeypatch, runtime_apps_response):
 
     plugin_config = config('plugins/ops/config-manager/config/plugin-config.json')
     result = await execute_event(app_config=plugin_config, event_name="runtime-apps-config", payload=None)
+
     assert result == runtime_apps_response
+
+
+@pytest.mark.asyncio
+async def test_runtime_apps_config_expand_events(monkeypatch, runtime_apps_response_exp):
+    app_config = config('apps/examples/simple-example/config/app-config.json')
+    monkeypatch.setattr(
+        runtime,
+        "server",
+        MockServer(app_config)
+    )
+
+    plugin_config = config('plugins/ops/config-manager/config/plugin-config.json')
+    result = await execute_event(
+        app_config=plugin_config, event_name="runtime-apps-config",
+        payload=None, expand_events=True
+    )
+
+    assert result == runtime_apps_response_exp